- fixes callbacks handling - fixes update contact - fixes thread model - working on send file

+++ app/src/main/java/ir/fanap/chat/sdk/application/chatexample/ChatContract.java
@@ -58,6 +58,6 @@ public interface ChatContract {
 
         void forwardMessage(long threadId, ArrayList<Long> messageIds);
 
-        void updateContact(long id, String firstName, String lastName, String cellphoneNumber, String email);
+        void updateContact(int id, String firstName, String lastName, String cellphoneNumber, String email);
     }
 }
